Rename CSharp to Dotnet everywhere
