+++ spec/dependabot/file_parsers/dotnet/nuget_spec.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
-require "dependabot/file_parsers/c_sharp/nuget"
+require "dependabot/file_parsers/dotnet/nuget"
 require_relative "../shared_examples_for_file_parsers"
 
-RSpec.describe Dependabot::FileParsers::CSharp::Nuget do
+RSpec.describe Dependabot::FileParsers::Dotnet::Nuget do
   it_behaves_like "a dependency file parser"
 end
