+++ spec/dependabot/file_updaters/dotnet/nuget_spec.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
-require "dependabot/file_updaters/c_sharp/nuget"
+require "dependabot/file_updaters/dotnet/nuget"
 require_relative "../shared_examples_for_file_updaters"
 
-RSpec.describe Dependabot::FileUpdaters::CSharp::Nuget do
+RSpec.describe Dependabot::FileUpdaters::Dotnet::Nuget do
   it_behaves_like "a dependency file updater"
 end
