+++ spec/dependabot/metadata_finders/dotnet/nuget_spec.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
-require "dependabot/metadata_finders/c_sharp/nuget"
+require "dependabot/metadata_finders/dotnet/nuget"
 require_relative "../shared_examples_for_metadata_finders"
 
-RSpec.describe Dependabot::MetadataFinders::CSharp::Nuget do
+RSpec.describe Dependabot::MetadataFinders::Dotnet::Nuget do
   it_behaves_like "a dependency metadata finder"
 end
