+++ spec/dependabot/update_checkers/dotnet/nuget_spec.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
-require "dependabot/update_checkers/c_sharp/nuget"
+require "dependabot/update_checkers/dotnet/nuget"
 require_relative "../shared_examples_for_update_checkers"
 
-RSpec.describe Dependabot::UpdateCheckers::CSharp::Nuget do
+RSpec.describe Dependabot::UpdateCheckers::Dotnet::Nuget do
   it_behaves_like "an update checker"
 end
